Create casting2.java
Let us look at the the following code. We are adding an int to a double. We know that an int is less precise than a double. In the line of code where we say double y=x+2 java will convert 2 to 2.0 and add it to 1.1 and this will print 3.1. There was casting that took place.
Casting hierarchy: //byte > short > int > long > float > double